requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
gunicorn>=23.0.0
playwright>=1.52.0
boto3>=1.38.0
python-dotenv>=1.1.0
pydantic>=2.11.0
python-multipart>=0.0.20
